Request with product image

+++ shopapp/src/main/java/com/project/shopapp/services/CategoryService.java
@@ -1,8 +1,11 @@
 package com.project.shopapp.services;
 
+import com.project.shopapp.components.LocalizationUtils;
 import com.project.shopapp.dtos.CategoryDTO;
+import com.project.shopapp.exceptions.DataNotFoundException;
 import com.project.shopapp.models.Category;
 import com.project.shopapp.repositories.CategoryRepository;
+import com.project.shopapp.utils.MessageKeys;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -10,8 +13,9 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class CategoryService implements ICategoryService{
+public class CategoryService implements ICategoryService {
     private final CategoryRepository categoryRepository;
+    private final LocalizationUtils localizationUtils;
 
     @Override
     public Category createCategory(CategoryDTO category) {
@@ -23,9 +27,10 @@ public class CategoryService implements ICategoryService{
     }
 
     @Override
-    public Category getCategoryById(Long id) {
+    public Category getCategoryById(Long id) throws DataNotFoundException {
         return categoryRepository.findById(id)
-                .orElseThrow( () -> new RuntimeException("Category not found"));
+                .orElseThrow(() -> new DataNotFoundException(localizationUtils
+                        .getLocalizedMessage(MessageKeys.CATEGORY_NOT_FOUND)));
     }
 
     @Override
@@ -34,14 +39,20 @@ public class CategoryService implements ICategoryService{
     }
 
     @Override
-    public Category updateCategory(Long categoryId, CategoryDTO category) {
-        Category existingCategory = getCategoryById(categoryId);
+    public Category updateCategory(Long categoryId, CategoryDTO category) throws DataNotFoundException {
+        Category existingCategory = categoryRepository.findById(categoryId)
+                .orElseThrow(() -> new DataNotFoundException(localizationUtils
+                        .getLocalizedMessage(MessageKeys.CATEGORY_NOT_FOUND)));
         existingCategory.setName(category.getName());
         return categoryRepository.save(existingCategory);
     }
 
     @Override
-    public void deleteCategory(Long id) {
-        categoryRepository.deleteById(id);
+    public void deleteCategory(Long id) throws DataNotFoundException {
+        if (categoryRepository.existsById(id)) {
+            categoryRepository.deleteById(id);
+            return;
+        }
+        throw new DataNotFoundException(localizationUtils.getLocalizedMessage(MessageKeys.CATEGORY_NOT_FOUND));
     }
 }
